完成Solution Question Distance

+++ src/main/java/com/may/routeplansystem/entity/po/Solution.java
@@ -17,5 +17,5 @@ public class Solution {
     private double totalDis;
     private double totalTime;
     private int version;
-    private byte delFlag;
+    private int delFlag;
 }
